Dessi_galli-main/requirements.txt
asgiref==3.5.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djangorestframework-yaml==2.0.0
Pillow==9.0.1
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tzdata==2021.5
